app/test: add SoC infra probe/detach test
The test probes for devices and tries to match with the test_driver0. The
test_driver0 assumes a fake device. The fake device is created by the provided
script soc_test_init.sh which creates a fake sysfs and fdt in the current
directory. Execute the test by:

  $ ./soc_test_init.sh setup
  $ export SYSFS_SOC_DEVICES=`pwd`/sys/bus/platform/devices
  $ expott FDT_ROOT=`pwd`/device-tree
  $ build/app/test --no-pci --no-huge --no-hpet -c 0x1 -n 1
  $ ./soc_test_init.sh clean

I did not find a proper way how to set the compatible property (a const char *)
list for the driver as DPDK has very strict compilation settings. I have no
idea how to setup the compatible with the const string unless the following
option is set:

 -Wno-error=incompatible-pointer-types